Introduce new component to render condition list with toggles to reduce duplication
